Reuse the current page if the HAR size is reasonably small.

+++ server/proxy_util.py
@@ -4,10 +4,20 @@ import datetime
 import logging
 import time
 
+import dateutil.parser
+import dateutil.tz
 import requests
 
 
+# Use this for creating the minimum datetime.
+# Never use datetime.datetime.min, as it's not aware of timezones.
+DATETIME_MIN = datetime.datetime(
+    datetime.MINYEAR, 1, 1, tzinfo=dateutil.tz.tzutc())
+
+
 class HarManager(object):
+
+    MAX_HAR_SIZE = 100 * 1024
 
     def __init__(self, args):
         self._logger = logging.getLogger('kcaa.proxy_util')
@@ -20,6 +30,10 @@ class HarManager(object):
 
     def reset_proxy(self):
         self.pageref = 1
+        self.next_pageref = None
+        self.old_pagerefs = set()
+        self.last_page_size = 0
+        self.last_accessed_time = DATETIME_MIN
         # At the initial trial, the proxy controller may not be ready.
         last_error = None
         for _ in xrange(10):
@@ -32,7 +46,7 @@ class HarManager(object):
                 time.sleep(1.0)
         else:
             raise last_error
-        if r.status_code != requests.codes.not_found:
+        if r.status_code != requests.codes.NOT_FOUND:
             r.raise_for_status()
         r = requests.post(self.proxy_root, data={'port': self.proxy_port})
         r.raise_for_status()
@@ -40,32 +54,42 @@ class HarManager(object):
                                                'captureContent': 'true'})
         r.raise_for_status()
 
-    def get_next_page(self):
+    def delete_old_pages(self):
+        if not self.old_pagerefs:
+            return
+        old_pagerefs = ','.join(map(str, self.old_pagerefs))
+        r = requests.delete('{}/{}'.format(self.proxy_har_pageref,
+                                           old_pagerefs))
+        # We can safely ignore errors, because old pages don't affect the
+        # behavior of get_current_page(). Just expect this will succeed
+        # sometime in the future.
+        if r.status_code == requests.codes.OK:
+            self.old_pagerefs.clear()
+        else:
+            self._logger.info('Failed to delete pages {}.'.format(
+                old_pagerefs))
+
+    def get_current_page(self):
         # TODO: BrowserMob Proxy seems a single threaded; if a proxy request
         # from the browser takes time. I.e. this method could take a long time
         # if it encounter such a situation. Use timeout and pageRefs wisely to
         # avoid that, if it's really required.
         start = datetime.datetime.now()
-        next_pageref = self.pageref + 1
-        rp = requests.put(self.proxy_har_pageref,
-                          data={'pageRef': next_pageref})
-        rp.raise_for_status()
-        rg = requests.get('{}?pageRef={}'.format(self.proxy_har,
-                                                 self.pageref))
-        rg.raise_for_status()
-        rd = requests.delete('{}/{}'.format(self.proxy_har_pageref,
-                                            self.pageref))
-        rd.raise_for_status()
-        self.pageref = next_pageref
+        r = requests.get('{}?pageRef={}'.format(self.proxy_har, self.pageref))
+        if r.status_code != requests.codes.OK:
+            self._logger.warn('Failed to get page {}.'.format(self.pageref))
+            return None, self.last_page_size
         fetch_span = datetime.datetime.now() - start
 
         # No Content-Length header?
-        content_size = len(rg.text)
+        content_size = len(r.content)
         # HAR content should always be encoded in UTF-8, according to the spec.
         start = datetime.datetime.now()
-        har = rg.json(encoding='utf8')
+        har = r.json(encoding='utf8')
         parse_span = datetime.datetime.now() - start
-        if (content_size >= 102400 or fetch_span.total_seconds() > 0.5 or
+
+        if (content_size >= HarManager.MAX_HAR_SIZE or
+                fetch_span.total_seconds() > 0.5 or
                 parse_span.total_seconds() > 0.5):
             self._logger.debug('HAR page {} ({:.1f} KiB), fetched in {:.2f} '
                                'sec, parsed in {:.2f} sec.'.format(
@@ -73,4 +97,49 @@ class HarManager(object):
                                    (1.0 / 1024) * content_size,
                                    fetch_span.total_seconds(),
                                    parse_span.total_seconds()))
-        return har
+
+        return har, content_size
+
+    def get_current_page_and_create_next(self):
+        # If there is no next page created yet, create it first.
+        if self.next_pageref is None:
+            next_pageref = self.pageref + 1
+            r = requests.put(self.proxy_har_pageref,
+                             data={'pageRef': next_pageref})
+            if r.status_code != requests.codes.OK:
+                self._logger.warn('Failed to create the next page {}.'.format(
+                    next_pageref))
+                return None, self.last_page_size
+            # Set next_pageref to mark that the next page is created.
+            self.next_pageref = next_pageref
+        har, content_size = self.get_current_page()
+        if not har:
+            return None, self.last_page_size
+        # If we get the whole content in the current page, we can move on to
+        # the next page. The last page can be deleted.
+        self.old_pagerefs.add(self.pageref)
+        self.pageref = self.next_pageref
+        self.next_pageref = None
+        return har, 0
+
+    def get_new_entries(self, har):
+        entries = []
+        last_accessed_time = DATETIME_MIN
+        for entry in har['log']['entries']:
+            started_datetime = dateutil.parser.parse(entry['startedDateTime'])
+            if started_datetime <= self.last_accessed_time:
+                continue
+            if started_datetime > last_accessed_time:
+                last_accessed_time = started_datetime
+            entries.append(entry)
+        if self.last_accessed_time < last_accessed_time:
+            self.last_accessed_time = last_accessed_time
+        return entries
+
+    def get_updated_entries(self):
+        self.delete_old_pages()
+        if self.last_page_size >= HarManager.MAX_HAR_SIZE:
+            har, self.last_page_size = self.get_current_page_and_create_next()
+        else:
+            har, self.last_page_size = self.get_current_page()
+        return self.get_new_entries(har) if har else None
